requirements.txt
django==4.2.6
djangorestframework==3.14.0
django-rest-swagger==2.2.0
drf_yasg==1.21.7
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
google-auth==2.23.3
psycopg2-binary==2.9.6
gunicorn==21.2.0
beautifulsoup4==4.12.2